Remove weight check for multiply commute (#4606)
Commute of multiply through conv2d is currently only allowed when both
weight and scale are constants. This is too restrictive and permits
fusing in case weight is not coming directly from block argument, but
through some intermediate op which can be const-eval.

We will remove this constraint since its not necessary and blocks
multiply commute in case we have bfp8 types in graph.