Add L1 interleaved fallback analysis to Optimizer pass (#4185)
### Problem description
Previously, only the costly `MemoryLayoutAnalysis` (when enabled)
attempted to change the memory configuration of DRAM interleaved ops to
L1 sharded, which requires complex compatibility checks and trade-offs.
As a result, DRAM-resident ops were not further optimized for L1
interleaved memory layout, missing a simpler opportunity to improve
memory locality and performance. The existing analysis focuses on legal
layouts and reconfiguration, but there was no dedicated pass to
selectively convert DRAM ops to L1 interleaved when beneficial,
especially with respect to L1 usage and scheduling constraints.

### What's changed
This PR adds a new analysis and transformation inside the `Optimizer`
pass that runs after the optional `MemoryLayoutAnalysis` and all its
scheduling and graph changes are complete. The pass traverses the
scheduled ops, identifies DRAM-resident ops, and attempts to convert
them to L1 interleaved layout if:
1. The op's inputs and outputs together fit in available L1 memory.
2. The op has only one use, and that use (consumer) is the immediate
next op in the schedule.
3. If user op has its output ~~already in L1~~ in either L1 or DRAM,
check if that op's inputs and outputs together fit in available L1
memory with this changed input.

The analysis uses L1 usage accounting similar to `ShardSolver` but
independently of it. For ops that meet the criteria, the pass changes
the op's output layout to L1 interleaved. For now, it does not insert
`ToLayout` ops or attempt further layout adjustments. Ops that do not
have exactly one use or whose consumer is not immediately scheduled
after them, as well as those affected by the row-major workaround which
currently follows this analysis in the Optimizer pass, are left
unchanged to avoid unnecessary or unsafe layout changes.

This targeted approach improves memory locality for eligible ops,
potentially reducing DRAM traffic and improving performance, while
maintaining correctness and respecting scheduling constraints.

There are a few simple tests for this using llvm-lit and FileCheck, and
it was separately successfully ran on the Resnet and few other benchmark
model but it still needs to be tested on larger models with more ops.
This feature is by default disabled, flag
`l1-interleaved-fallback-analysis-enabled` is toggled independently of
`memory-layout-analysis-enabled`.

### Checklist
- [x] New/Existing tests provide coverage for changes
- [x] Flag for enabling the feature (default:`
l1-interleaved-fallback-analysis-enabled`=false)