[D2M] Add Support for DRAM Interleaved Buffers in D2M Runtime (#4561)
### Ticket
N/A

### Problem description
The TTMetal TTRT backend does not currently support DRAM interleaved
buffers, and all DRAM buffers are assumed to be sharded. A partner wants
to use DRAM interleaved buffers, and D2M will need to eventually support
these buffers even if they are not the first choice.

### What's changed
- D2M memrefs are assumed to be DRAM interleaved if:
  - the memory space is set to DRAM, and
  - there is no `ShardLayoutAttr` (tt/ttcore.shard in the IR)
- For each DRAM interleaved buffer, create and serialize an
`InterleavedBufferConfig` instead of a `ShardedBufferConfig` in
`TTMetalToFlatbuffer`
- Modified the TTMetal flatbuffer schema to allow exactly one of
`ShardedBufferConfig` or `InterleavedBufferConfig` for each serialized
buffer
- Modified the TTMetal TTRT backend to check the type of `bufferConfig`
in each buffer and call the correct `CreateMeshBuffer` with the correct
config object type.

Note: Interleaved buffers must be single device and in DRAM. L1
interleaved or multi-device DRAM interleaved buffers are not supported

Note: **This PR does not add a ttir->TTMetal lowering for DRAM
interleaved tensors.** The partner has their own lowering pathway into
TTMetal

### Testing
Manually tested. CI testing to be discussed in D2M F2F as it would
require checking in a TTMetal IR file

### Checklist
- [ ] New/Existing tests provide coverage for changes

---------

Co-authored-by: Brett Grady <bgrady@tenstorrent.com>