[tt-explorer] Add cli flag for enable/disable model execution (#4677)
### Ticket
N/A

### Problem description
We need a way to control if model execution is enabled or disabled for
`tt-explorer`.

This will enable easy hosting of the tool while hiding the more resource
intensive parts of it.

### What's changed
Added a flag to `tt-explorer` for controling if model execution is
enabled or not.

### Checklist
- [x] New/Existing tests provide coverage for changes

Is blocked by: https://github.com/tenstorrent/model-explorer/pull/65