Fix TM count bug in EIO, apply FlattenSlidingWindow to `ConvTranspose2dOp` (#4667)
### Tickets
https://github.com/tenstorrent/tt-mlir/issues/4654,
https://github.com/tenstorrent/tt-mlir/issues/4637

### Problem description
The `countTMs` function in EraseInverseOps.cpp always counts 0 TMs, this
is not intended.
`ConvTranspose2dOp` does not have `FlattenSlidingWindow` compatibility
which can hinder EIO in erasing inverse ops.

### What's changed
Add `FlattenedCompatInfoAttr` as an argument to `ttir.conv_transpose2d`
and include functionality in `FlattenSlidingWindow` to flatten it.

Fix the trait check in `countTMs`

### Checklist
- [x] New/Existing tests provide coverage for changes

---------

Co-authored-by: mvasiljevicTT <mvasiljevic@tenstorrent.com>