[D2M] add optional affine map to MetalLayoutAttr (#4703)
### Ticket
NA

### Problem description
Currently, we have a view_layout op, but in tensor-space it does not
include an affine map. Instead, the affine map is computed during
bufferization and added to ViewLayoutAttr in memref-space. This works
for simple reblocking, but does not work for things like tranpose.

### What's changed
Add a field in tensor-space MetalLayoutAttr to describe affine map. For
something like TransposeOp, we will set this by hand to ensure we get
the correct values in future PRs. Added manual IR check to ensure this
gets properly converted to ViewLayoutAttr during bufferization.

### Checklist
- [X] New/Existing tests provide coverage for changes