Enable convolution bias fusing when input is broadcasted (#4732)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-mlir/issues/4659)

### Problem description
In unet from torch-xla, the transposed convolution bias appears as an
add op whose input is already broadcasted to the output shape. Because
of this, the current implementation cannot fuse it with the
`convolutionOp`.

### What's changed
The `ConvWithBias` pattern now checks if the add input comes from a
broadcast. If so, it uses the broadcast’s input directly as the bias.

### Checklist
- [x] New/Existing tests provide coverage for changes