Mention `TTMLIR_TOOLCHAIN_DIR` instead of hardcoding a path (#4769)
### What's changed

If one wants to use a different path, copy-pasting the current
instructions is daunting because they need to change the path in 3
different places, consistently. Referring to the variable defined at the
beginning of the instructions makes this process much smoother instead
(they need to change the path only in one place).

### Checklist
- [ ] New/Existing tests provide coverage for changes