[Optimizer] Add constraint APIs for device related ops (#4774)
### Ticket
#4207 

### Problem description
Constraint APIs for device related ops

### What's changed
Constraint APIs are not needed for device related ops, but they block
the end goal of adding the APIs to `TTNNOps` class. Therefore, we just
add the APIs and issue errors.
- `getDeviceOp`: is used once in each input graph and the optimizer does
not need to make modifications to it.
- `toDeviceOp` and `fromDeviceOp`: require memory IO, cannot be
implemented.
- `ToDTypeOp `: host only op, cannot be implemented rn.

Some info can be found
[here](https://github.com/tenstorrent/tt-mlir/issues/4207#issuecomment-3224551014).

### Checklist
- [X] Existing tests provide coverage for changes