Add placeholder constraints API for scatter (#4789)
### Ticket
#4196

Calling `getOpConstraints()` or `getOpRuntime()` for `scatter` will now
trigger `issueError()`.

As mentioned in issue #4788 , `scatter` is currently defined as a binary
composite op in TTNN IR, with a workaround implementation in runtime.
This should be fixed soon, and then we can update the constraints API to
properly call `ttnn::scatter`.