Enable BatchNorm decomposition only with fusion flag on (#4796)
### Problem description
BatchNorm is currently always decomposed after a `ConvolutionOp`, but
fusion with convolution happens only if the
`enable-fusing-conv2d-with-multiply-pattern` flag is enabled.
When the flag is disabled, BatchNorm still gets decomposed but not
fused, leading to a performance drop.

### What's changed
BatchNorm is now decomposed only if the
`enable-fusing-conv2d-with-multiply-pattern` flag is enabled.

### Checklist
- [ ] New/Existing tests provide coverage for changes