[CI] Extend timeout (#4840)
### Ticket

### Problem description
Some workflows fail due to timeout especially during peak hours when
network lags due to the traffic and adds significant time to the jobs.

### What's changed
Quick fix is to set timeout to 70 minutes which will not make huge
difference.
Safe timeout is considered to be 2 times normal test duration and should
be triggered only in case of hangs/machine failures/etc.

### Checklist
- [ ] New/Existing tests provide coverage for changes