D2M: support non-tile-aligned tensor shapes (#4709)
### Ticket
Related-To: #3013 #3014 #3015 #3023
Supersedes: #3805

### Problem description
Currently when we pass a tensor from the runtime, its elements are
assumed to be densely packed into a flat row-major storage.

This means that, the elements of a tensor with a non-tile-aligned shape
end up in completely wrong locations in the DRAM and thus in the CB.

Additionally, due to the grid & tile alignment logic (ref. #4131),
non-tile-aligned tensors will have strictly smaller storage than their
device counterparts, overflow/segfault is thus ensured during
`Enqueue{Read,Write}Buffer`.

Functionally this is one of the steps to unblock the work on tile-level
broadcast and some other ops, as zero-padding is required in the last
two dimensions if they are not already tile-aligned.

### What's changed

- The calculation and application of alignments in `MetalLayoutAttr` is
updated to always at least align to tiles, among other minor changes.
- In the bufferization pass, when unaligned host IO is detected, a host
bounce buffer with fully-aligned size is created to do I/O with the
device buffer, and a `memref.copy` between the bounce buffer and the
tightly-packed host buffer will be issued to copy the elements over.
- Create a new `TTCore_HostLayoutAttr` attribute to pass (potentially
aligned) tensor volume & strides info to the host runtime, attached to
the `memref` associated with the bounce buffer.
- The `BufferDesc` struct in metal flatbuffers now has the new
`host_strides` and `host_volume` fields to store the corresponding
values from the `HostLayoutAttr`.
- The `TensorDesc` struct in the metal runtime uses its existing
`stride` field to store the (potentially aligned) strides read from the
flatbuffer, as well as a new `physicalVolume` field to store the
`host_volume` value.
- In the executor, `HostAllocCommand` now allocates zero-filled tensor
storage with the requested physical volume.
- The extra `memref.copy`, which eventually gets lowered to
`ttmetal::memcpy`, now does strided memcpy according to the operand's
`host_strides`, if any of the two operands is aligned & padded.
- Add golden tests that verify the addition of non-tile-aligned shapes
from 2D to 5D, and various other tests.

### Known issues

- The bounce buffers respect both `8x8` grid alignment and `32x32` tile
alignment, and thus could lead to extreme waste for shapes like
`257x1x1x1`.
- Pad value is always 0 (by using `callocShared`), but this might be
wrong from some reduction ops (max/min/prod).
- The alignment is fixed, while we need to support custom alignments
#4852.

### Checklist
- [x] New/Existing tests provide coverage for changes

---------

Signed-off-by: wenbinlyuTT <wenbinlyu@tenstorrent.com>