Initial support of trace in EmitC (#4773)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3672

### Problem description
There isn't a way to codegen trace in EmitC.

### What's changed
Added a lowering for trace-related ops in TTNN to EmitC. Most of them
are straightforward lowering for ops that exist in TTNN lib. The only op
that required non-trivial lowering logic is
`ttnn.capture_or_execute_trace_op`. It's implemented as a call to the
generated function, which on first call calls `capture_calle`, and
afterwards `execute_calle`. State between calls is stored in global
variables. Initially, I planned to track in the state through local
variables with static lifetime, but EmitC doesn't support static
variables.

Since there isn't `TraceId` as a type in TTNN dialect (it's modeled as a
tensor with a single scalar value), I've added a `ttnn.trace_id`
attribute that's stored in the encoding of the tensor for the tensor
that represents `TraceId`. This is necessary because otherwise,
`ttnn::Tensor` would be generated as a type for the trace id.

I've tested in manually through ttnn-standalone, but there isn't an easy
way currently to test through ttrt (that's the same issue for trace
through the runtime). I'll look into alternatives to enable testing in
CI (https://github.com/tenstorrent/tt-mlir/issues/4772).

### Checklist
- [ ] New/Existing tests provide coverage for changes