Fixing analyze mesh pass to also take round trip shardy into account   (#4900)
With the uplift to jax 0.7.1, we have deleted the round trip pass from
tt-xla (which converts round trip shardy into normal shardy dialect), as
jax to longer emitted it. However, torch_xla still emits it in that
shape, so the AnalyzeMesh pass needs to take that into account. In this
case, that means dealing with the `xla.sdy.meshes` that comes from
torch_xla.