Add `ttnn.load_tensor` and `ttnn.dump_tensor` (#4998)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/4965

### Problem description
https://github.com/tenstorrent/tt-mlir/pull/4909 has added a runtime
support for these operations, but we need to model it on the TTNN
dialect level. We have a use case for `load_tensor` in codegen paths.

### What's changed
Added support for these ops on the dialect level. In the runtime, I've
added an implementation for ops execution that's unrelated to
https://github.com/tenstorrent/tt-mlir/pull/4909, since these are pretty
much direct calls to the `tt-metal` APIs.

### Checklist
- [x] New/Existing tests provide coverage for changes