[CI] Refurbish tt-mlir workflow part 1 (#4968)
### Ticket

### Problem description
CI needs optimization.

### What's changed
Selective archive of install and build artifacts.
This will lower the size of install for ~400MB and build for ~1GB.
At the same time, it will speed up build for about ~2min and (each) test
for ~5m.
This can lower the machine utilization and decrease OnPR time for
~30min.
pykernel test duration lowered from ~30min (26min for test) to 4.5min
(15s for test).
All timings are estimations based on average usage. 

### Checklist
- [ ] New/Existing tests provide coverage for changes