D2M: use SFPU for add/sub/mul (#5029)
### Ticket
None

### Problem description
The FPU add/sub/mul ops are reading from the CBs and writing to the DST,
makes it difficult/costly to fuse them even with other eltwise ops,
blocking our op fusion & bcast work.

### What's changed
Switch to the SFPU variants, the FPU ones are still present in the
TTKernel dialect, but are orphans now.
Fix an existing bug that leads to incorrect SSA ordering when the DST
index is loop-dependent.

### Known issue
The comparison ops are decomposed into sub+cmpz, and will have two
`init_sfpu` inserted. This is localized in the D2M comparison ops and is
benign atm. The duplicates will be removed by the soon-to-land init
hoisting work.

### Checklist
- [x] New/Existing tests provide coverage for changes

---------

Signed-off-by: wenbinlyuTT <wenbinlyu@tenstorrent.com>