[D2M/TTNN Integration] Add TTNN Layout Translation Logic in TTIRToTTIRGeneric and Translate Block Sharded TTNN Tensors to Metal Tensors (#5052)
### Ticket
#5050 and #5019 

### Problem description
As part of D2M/TTNN integration, the D2M pipeline will be invoked with
ttir ops that have tensor operands with `TTNNLayoutAttr` (i.e. tensors
whose layout is set by TTNN). The D2M pipeline needs to:
1. parse these tensors to run existing D2M passes, and
2. maintain the tensor layouts as set by TTNN to ensure D2M-generated
ops and downstream TTNN ops access the tensor in the same way

### What's changed
- Modified the layout generation logic in the `TTIRToTTIRGeneric` pass.
When invoked with the `ttnn-mode=true` option, this pass:
  1. matches ttir ops with TTNN tensor operands and/or results
2. creates an equivalent metal tensor with fully specified grid/shard
shape and `MetalLayout` for each TTNN tensor
3. inserts a `ttir.ttnn_ttmetal_cast_op` to represent the layout
translation
a. The casts are inserted `ttnn tensor -> metal tensor` for op operands
and `metal tensor -> ttnn tensor` for op results
- Added a new `ttnn-mode` option to `TTIRToTTIRGeneric` and connected it
to the `ttnn-mode` pipeline option in `ttir-to-ttmetal-pipeline`. When
this option is false, the pass asserts when encountering TTNN tensors as
the behaviour is undefined
- Currently only block sharded L1 tensors are supported, with more to be
added
- Added IR tests, including testing integration with
https://github.com/tenstorrent/tt-mlir/pull/4978

### Checklist
- [X] New/Existing tests provide coverage for changes

---------

Co-authored-by: Vincent Tang <vtang@tenstorrent.com>