Ops with # threads <= # DMA HW Threads don't need to be merged
Previously we were making the assumption that there will always be a 1-1
relationship between the number of operands and the number of
datamovement regions.  In the context of manually written generic ops,
this is not necessarily true, and the datamovement regions are already
manually fused.

Instead we should count the number of datamovement threads which is more
accurate.