TTKernel to EmitC ArithFloorDivRewriter (#5115)
Arith FloorDivSIOp doesn't have an emitc lowering, probably because of
the spec which says:

> Signed integer division. Rounds towards negative infinity, i.e. 5 / -2
= -3

However we know our index type will map to size_t which is unsigned,
making a negative denominator impossible, so as long as we assert that
this floordiv is working on values of `index` type it's safe to map this
op to regular divi.

Prerequisite for DSL 2 work.