Amend workaround for update cache op to extract a replicated updateIndex (#5022)
Due to #1510, the update_cache op's updateIndex tensor must be returned
to host and dissected to provide a uint32_t to TTNN.

This workaround does not work for a distributed tensor - for tensor
parallel static cache update, the updateIndex may end up replicated
across multiple devices, and ::ttnn::get_host_buffer explicitly asserts
against this case.

This change uses the MeshComposer API to coalesce the replicated
updateIndex into a single tensor and returns it to host. This continues
to work if the updateIndex is not replicated.