[Explorer] Make `ttrt` dependency dynamically load (#5143)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/4614

### Problem description
For hosting `tt-explorer` we would not have access to specific hardware,
so we want to do two things:
1. Make explorer execute without support for execution (that would also
help with security)
2. Make explorer dependency on `ttrt`, which depends on specific
hardware, optional

### What's changed
This PR helps solving item number 2 from above by making the dependency
on `ttrt` not required.

It works by checking for `ttrt` module availability and importing it if
it is available.

It also provides better support for disabling execution by making the
cli flag take precedence over the actual module existence.

### Checklist
- [ ] New/Existing tests provide coverage for changes