[runtime] `hasLayout` API (#5157)
Introducing function `bool hasLayout(Tensor t, Layout l)` to the runtime
API. This function returns whether the provided tensor has the specified
layout.

This is needed since the `toLayout()` function has a side-effect
(dirtying of the tensor) regardless of whether or not the tensor is
already in the correct (desired) layout - which affects the tensor cache
for consteval. So, in FEs we would like to be able to know if the
`toLayout()` is needed before we execute it - e.g. in scenarios where
one tensor is used with multiple programs with different layouts.

Issue tenstorrent/tt-xla#953