Create bypass path for ScatterOp in UpdateGlobalToLocalShapesPass when applied to sharded static cache update (#5084)
In commit `dad89819`, an assert was added to block stablehlo.scatter
from lowering in UpdateGlobalToLocalShapesPass, causing a regression for
TP models with static cache, because a static cache update lowers to a
scatterOp.

Instead of preventing scatter from being lowered in all cases, add a
narrow bypass path for one case of a "safe" scatter where the op's attrs
do not need to be updated. This is done by detecting a cache update
semantic and checking if the "scatter axis" is orthogonal to the
input/update sharding axis, which matches a Transformers static cache
update (torch.index_copy_) when lowered to stablehlo through tt-xla and
tt-torch.

This unblocks tensor parallel static cached LLM compilation in tt-xla.

A longterm, general solution is to be made in #3496