Update the stream_layout op's verifier
The stream layout previously didn't have a stringent enough verifier,
which is now able to validate that shard shapes between the storage and
the result type are equal.

There was one bit of fallout that was detected by an existing test
`test/ttmlir/Dialect/D2M/allocate/generic_form_streams_matmul.mlir`.
The subtest `matmul_multi_core_transpose` uses a view_layout op to
transpose the input on the fly.  The allocator needed a small update to
the stream_layout's storage memref type which must come from the post
view's shape.  This is accomplished by reading the generic's operand
type.