D2M to TTKernel SFPU op placement fix
Previously we were attempting to hoist the SFPU op to its earliest
possible insertion point by analyzing its relative location to its
operands.  This breaks down when the operands span blocks, we can only
safely apply this movement if we can guarantee we're pushing the op
downward in program order and in the same block.