Implement TTNN to TTIR for Matmul op (#5167)
### Ticket
Closes #5158

### Problem description
The `ttnn.matmul` op is not supported in `ConvertTTNNToTTIR`.

### What's changed
Extends the `ConvertTTNNToTTIR` pass to support converting
`TTNN_MatmulOp` to `TTIR_MatmulOp`. The approach is the same as for
elementwise ops. TTNN attributes not present in TTIR are _not_
preserved.

The pass is updated to make untagged TTNN operations dynamically legal
and do a full conversion.