[TTNN/D2M Integration] Erase Inverse TTNNMetalLayoutCast Ops (#5181)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/5178

### Problem description
`TTNNMetalLayoutCastOp` is a representational operation that allows a
`RankedTensorType` object to be seen with either a `MetalLayoutAttr` or
a `TTNNLayoutAttr`. In TTNN JIT and TTNN/D2M integration flows, this op
is inserted to capture the input and output tensors whose layout is set
by TTNN and must be understood by D2M.

If a chain of ops is ingested by TTNN JIT or TTNN/D2M integration, the
result of each op will be casted metal -> ttnn and then again ttnn ->
metal for the input the consumer op. This is redundant and complicates
other optimizations. Inverse casts should be erased.

### What's changed
- Add a canonicalization pattern to eliminate cast ops whose input is
produced by another cast op
- Closes https://github.com/tenstorrent/tt-mlir/issues/5178

### Checklist
- [X] New/Existing tests provide coverage for changes