[EmitC/EmitPy] Load input tensors (#5037)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/5003

### Problem description
With the addition of `ttnn.load_tensor`, it's now possible to load
tensors from disks. We can use this in EmitC (and EmitPy) pipelines to
load input tensors instead of creating fixed `ttnn.ones`.

### What's changed
Added a pipeline option `load-input-tensors-from-disk` that will
generate a function that will load tensors with `ttnn.load_tensor`.
Currently, it's a boolean option; if specified, it's expected that all
inputs will be loaded and their names are hardcoded to `arg0.tensorbin`,
`arg1.tensorbin`,..., `argN.tensorbin`. These restrictions could be
weakened in the future if there is a necessity (for example, loading
some inputs from defined paths, while generating synthetic tensors for
others).
 
### Checklist
- [x] New/Existing tests provide coverage for changes