D2M: unify DST capacity calculation and take 32b mode into account (#5206)
### Ticket
Closes: #5023
Related-To: #4964 #5108 #5151

### Problem description
By default we set DST to `SyncHalf`, so after `fp32DestAccum` is enabled
in #4964, we're only left with 4 tiles to work with if we're doing f32
operations.

Existing code isn't aware of the DST's dynamic capacity (the fusion /
loop tiling / insert DST access passes), they assume 8 tiles and go
straight to overflow land.

Additionally, we only consider if the output can fit in the DST,
ignoring inputs and/or intermediate results, worsening the overflow
issue.

We typically only test small f32 tensors so this has rarely caused any
trouble in the CI. For the few tests that does trigger overflow, the
only symptom is PCC error and we simply masked the off (e.g. the matmul
tests and the unaligned shapes tests).

### What's changed
* Make clear distinction between the DST's physical size and logical
size in the code base.
* Unified logical DST capacity calculation logic as a helper function
under `ttcore::ChipDescAttr`.
* Now the logical capacity of the DST will be cut in half if any of the
operands of a GenericOp is 32b.
* Renamed `max-dst-register-size-tiles` to `max-dst-physical-size-tiles`
and support this option in the insert DST access pass as well.
* Be explicit about our choice of setting `SyncHalf`, left a note about
its implications.
* Switch to conservative loop tiling if any of the compute op uses more
than one DST tile, or doesn't use the DST in-place.
* In the insert DST access pass, we now assert if we allocate more DST
slices than it can hold. The change above is the defense against this
situation.
* Memref linearization is disabled for block matmuls due to the
un-collapsible strided `memref.subview` resulted from limited DST
capacity
* Reworked the affected tests: original tests are switched to `f16/bf16`
to keep the shapes the same, the f32 variant of these tests are being
tested under the override ``max-dst-register-size-tiles=32`.
* Various cleanups, renames and comments to assist understanding.

### Checklist
- [x] New/Existing tests provide coverage for changes

---------

Signed-off-by: wenbinlyuTT <wenbinlyu@tenstorrent.com>