Implement isNotConflicting bufferization interface for dma (#5166)
This PR trivially implements `isNotConflicting` by unconditionally
returning true. Relevant snippet of background from the bufferizable op
interface:

> This method can be implemented to specify custom RaW analysis rules.
> If this method returns `true` the given OpOperands are not considered
> to be conflicting and do not force out-of-place bufferization.

Out of place bufferization is never desireable in the context of a
generic op region, it means that bufferization can legally introduce a
new `memref.alloc` to bounce a RaW conflict through. However, all memory
is preallocated for the generic and it's not legal to have a memref
alloc appear inside of the generic region.