Make the LLVM build optional during env setup (#5273)
Adds a new CMAKE flag `TTMLIR_BUILD_LLVM` which makes the LLVM build
step optional. This is useful for overriding with out-of-tree built LLVM
(for integration into third party projects or for building against a
debug version of LLVM, for example). When building with third party LLVM
I was not able to find `flatbuffers.h` - I believe this is because the
include path for the toolchain directory is not explicitly included and
LLVM includes were ensuring `flatbuffers.h` was found. I explicitly
added an include for the toolchain directory and the problem went away.

---------

Co-authored-by: Alex Baden <alex@kernelize.ai>